refactor: domain module di 및 기타 이동
- 기타사항 이동
