+++ app/build.gradle.kts
@@ -43,6 +43,8 @@ android {
 
 dependencies {
 
+    implementation(project(Project.domain))
+
     implementation(Dependencies.Androidx.core)
     implementation(Dependencies.Androidx.appcompat)
     implementation(Dependencies.Material.material)
