+++ data/src/main/java/com/capstone/artwhale/data/dto/AlbumDto.kt
@@ -1,8 +1,5 @@
 package com.capstone.artwhale.data.dto
 
-
-import com.capstone.artwhale.BuildConfig.BASE_URL
-import com.capstone.artwhale.domain.model.Album
 import com.google.gson.annotations.SerializedName
 
 data class AlbumDto(
@@ -22,12 +19,4 @@ data class AlbumDto(
     val title: String,
     @SerializedName("updatedAt")
     val updatedAt: String
-) {
-    fun toAlbum(): Album = Album(
-        id = id,
-        albumImgUrl = BASE_URL + fileId.path,
-        title = title,
-        mood = mood,
-        isLike = like,
-    )
-}
+)
