+++ data/src/main/java/com/capstone/artwhale/data/dto/LoginDto.kt
@@ -6,5 +6,3 @@ data class LoginDto(
     @SerializedName("email")
     val email: String
 )
-
-fun String.toLoginDto(): LoginDto = LoginDto(this)
