+++ data/src/main/java/com/capstone/artwhale/data/dto/MoodDto.kt
@@ -1,7 +1,6 @@
 package com.capstone.artwhale.data.dto
 
 
-import com.capstone.artwhale.domain.model.Mood
 import com.google.gson.annotations.SerializedName
 
 data class MoodDto(
@@ -9,9 +8,4 @@ data class MoodDto(
     val id: Int,
     @SerializedName("text")
     val text: String
-) {
-    fun toMood(): Mood = Mood(
-        id = id,
-        name = text,
-    )
-}
+)
