+++ data/src/main/java/com/capstone/artwhale/data/dto/MusicDto.kt
@@ -1,7 +1,5 @@
 package com.capstone.artwhale.data.dto
 
-import com.capstone.artwhale.BuildConfig.BASE_URL
-import com.capstone.artwhale.domain.model.Music
 import com.google.gson.annotations.SerializedName
 
 data class MusicDto(
@@ -25,15 +23,4 @@ data class MusicDto(
     val title: String,
     @SerializedName("updatedAt")
     val updatedAt: String
-) {
-    fun toMusic(): Music = Music(
-        id = id,
-        albumImgUrl = BASE_URL + albumArtId.fileId.path,
-        title = title,
-        singer = "singer",
-        time = duration,
-        isLike = like,
-        mood = mood,
-        musicUrl = BASE_URL + fileId.path,
-    )
-}
+)
