+++ data/src/main/java/com/capstone/artwhale/data/dto/UpdateNickNameDto.kt
@@ -6,5 +6,3 @@ data class UpdateNickNameDto(
     @SerializedName("nickname")
     val nickname: String
 )
-
-fun String.toUpdateNicknameDto() = UpdateNickNameDto(this)
