+++ domain/src/main/java/com/capstone/artwhale/domain/mapper/LoginMapper.kt
@@ -1,0 +1,5 @@
+package com.capstone.artwhale.domain.mapper
+
+import com.capstone.artwhale.data.dto.LoginDto
+
+fun String.toLoginDto(): LoginDto = LoginDto(this)
