+++ domain/src/main/java/com/capstone/artwhale/domain/mapper/RecentSearchMapper.kt
@@ -1,0 +1,8 @@
+package com.capstone.artwhale.domain.mapper
+
+import com.capstone.artwhale.data.dto.RecentSearchDto
+import com.capstone.artwhale.domain.model.RecentSearch
+
+fun RecentSearchDto.toRecentSearch(): RecentSearch = RecentSearch(keyword)
+
+fun RecentSearch.toRecentSearchDto() = RecentSearchDto(keyword)
