+++ domain/src/main/java/com/capstone/artwhale/domain/mapper/UpdateLikeMapper.kt
@@ -1,0 +1,5 @@
+package com.capstone.artwhale.domain.mapper
+
+import com.capstone.artwhale.data.dto.UpdateLikeDto
+
+fun Int.toUpdateLikeDto() = UpdateLikeDto(this)
