+++ domain/src/main/java/com/capstone/artwhale/domain/mapper/UpdateNickNameMapper.kt
@@ -1,0 +1,5 @@
+package com.capstone.artwhale.domain.mapper
+
+import com.capstone.artwhale.data.dto.UpdateNickNameDto
+
+fun String.toUpdateNicknameDto() = UpdateNickNameDto(this)
