+++ domain/src/main/java/com/capstone/artwhale/domain/repository/impl/MoodRepositoryImpl.kt
@@ -1,6 +1,7 @@
-package com.capstone.artwhale.data.repository
+package com.capstone.artwhale.domain.repository.impl
 
 import com.capstone.artwhale.data.datasource.mood.MoodDataSource
+import com.capstone.artwhale.domain.mapper.toMood
 import com.capstone.artwhale.domain.model.Mood
 import com.capstone.artwhale.domain.repository.MoodRepository
 import javax.inject.Inject
