+++ domain/src/main/java/com/capstone/artwhale/domain/repository/impl/NoticeRepositoryImpl.kt
@@ -1,6 +1,7 @@
-package com.capstone.artwhale.data.repository
+package com.capstone.artwhale.domain.repository.impl
 
 import com.capstone.artwhale.data.datasource.notice.NoticeDataSource
+import com.capstone.artwhale.domain.mapper.toNotice
 import com.capstone.artwhale.domain.model.Notice
 import com.capstone.artwhale.domain.repository.NoticeRepository
 import javax.inject.Inject
