+++ domain/src/main/java/com/capstone/artwhale/domain/repository/impl/UserRepositoryImpl.kt
@@ -1,7 +1,8 @@
-package com.capstone.artwhale.data.repository
+package com.capstone.artwhale.domain.repository.impl
 
 import com.capstone.artwhale.data.datasource.user.UserDataSource
-import com.capstone.artwhale.data.dto.toUpdateNicknameDto
+import com.capstone.artwhale.domain.mapper.toUpdateNicknameDto
+import com.capstone.artwhale.domain.mapper.toUserInfo
 import com.capstone.artwhale.domain.model.UserInfo
 import com.capstone.artwhale.domain.repository.UserRepository
 import com.capstone.artwhale.util.LocalPathUtil
